archive/samples/python/13.core-bot/requirements.txt
botbuilder-integration-aiohttp>=4.15.0
botbuilder-dialogs>=4.15.0
botbuilder-ai>=4.15.0
datatypes-date-time>=1.0.0.a2

